recomendation/requirements.txt
python-dotenv==1.0.1
psycopg2==2.9.9